requirements.txt
fastapi[all]
setuptools
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings